create-cloudfoundry: add "-experimental" deployment of paas-billing
this is cribbed from a reversion of
2251bef79486ee75f7094bd9b61098b9fd5b5125 and
4a444783479af11867fc46d0afede5e52e325b00. a parallel deployment of
billing in production will allow us to test experimental billing
code with real life billing data and ensure any performance
improvements we make are consistent.

pulling from the `experimental` branch on paas-billing, still requiring
a signed commit because we're working with live data.