billing-experimental deployment: use db named billing-experimental-db
this allows us to use a db that is shared from the billing space
without name collisions. the reason for using a space-shared
database is to allow us to start with a clone of the real production
`billing-db`, but clones can only easily be done into the same
space. so we have to choose a db name that won't collide with one
already in the production `billing` space.