Disable app autoscaler acceptance tests in dev
The app autoscaler acceptance tests often fail during dev pipeline runs.

We majoritively don't actually need these to run every time we run the
pipeline, as not an awful lot of what we do actually affects the
performance of the autoscaler.

To get this disabling working in a sensible way, I have introduced a new
`make` variable, `$DISABLE_APP_AUTOSCALER_ACCEPTANCE_TESTS`, which is
set to the value of `$SLIM_DEV_DEPLOYMENT` by default. This way, if
running a non-slim deployment, the tests will run, as the CF platform
should be better scaled to handle the extra load.

When generating the pipeline, if the new variable is not set, it will
default to the value of `$DISABLE_CF_ACCEPTANCE_TESTS`, ensuring that
that original variable will still function as expected.