Add a script to simplify suspending cf orgs
Allows suspending an org by name, and includes a 'hidden' unsupend
method.

This will make suspension simpler, as we won't have to copy and paste
from the docs or remember the syntax.

There is also a sanity check step to make sure you're suspending the
right org in the right environment.