log-caches: vertically scale rather than horizontally
we're currently seeing log-cache instances frequently having
>70% cpu usage while other instances are sitting near-idle.

at the same time, according to metrics, at any one time we only
appear to have ~10 app ids actively logging during any one
minute (london region). given that one app_id's logging is handled
by a single log-cache instance at any one time, this suggests
that attempting to horizontally scale much more than 10 instances
is not going to be successful in avoiding these high CPU spikes.
to do this we need to reduce the degree to which our compute
power is partitioned.

this PR doubles the size of log-cache instances and approximately
halves their number (if this change is successful we can go
on to reduce it further, possibly *under* half). the intention
being that any one log-cache should have twice the CPU available to
it (to allow it to handle chatty app_ids), but we should end up with
approximately the same amount of total memory for the log-caches
giving us a similar length of log retention.

the final cost should be similar because all r5 instances look the
same to our savings plans.