Change VPN configuration to meet supported Azure configuration
The NCSC profile
(https://www.ncsc.gov.uk/guidance/using-ipsec-protect-data)
suggests Diffie-Hellman groups 19 or 14 for both phases.

However, Azure's site-to-site VPN service does not support
those options. The matching groups are 14 for phase 1 and 2
for phase 2.

Equally AES128-GCM-16 is not supported by Azure, so we fall
back to AES128.