alerts: allow log-cache instances to reach 80% cpu util before warning
log-caches can surge to high utilization occasionally and it
appears to be "fine", at least to the point we would like
to quieten down this warning until we can figure out a proper
fix.

extend the test cases slightly to cover longer alert periods
and differentiate between >70% usage and >80% usage.