prometheus ops files: fixes for switch to firehose_exporter-attic
we added the symlink to `monitor-cf-attic.yml` too late - *after*
previous opsfiles had attempted to make changes to the
firehose_exporter job. these changes are discarded by the
`monitor-cf-attic.yml` opsfile when it is applied, so we need to
move customizations to after the renaming and target the new job
name, `firehose_exporter-attic`.