Unwrap the grafana dashboards
Previously, we stored the dashboards wrapped in the JSON object that
we use to send the dashboard to Grafana's API. This meant that when updating
dashboards, we have to wrap the dashboard from Grafana into the object
again before committing.

This commit unwraps the dashboard JSON that we are storing, and
actually creates the api call object within the pipeline.