waf: migrate to wafv2, adding ability to conditionally obey x-forwarded-for
wafv2 gives us the ability to use more complex sets of rules for
configuration of rate limiting.

firstly we exclude an environment's own egress IPs and "system"
(currently just concourse) IPs from rate limiting to prevent us
blocking either our own tests or tenants that are using externally
routed inter-app requests.

we then have a rate limit rule for requests we've determined should
*not* have x-forwarded-for respected. we determine this by looking
for a header with a "secret" value only set by our blessed CDNs. so
that the header is not visible to tenant apps we abuse the header
`x-cf-instanceid` for this. this is because neither load balancers
or the gorouter have the ability to remove arbitrary headers from
forwarded requests, so instead we have to choose one that we know
will be blindly overwritten by the gorouter.

finally there is a similar rate limit rule for requests that
*should* have x-forwarded-for respected. these are expressed as
two separate rules with the same header-checking condition inverted
because abort-early/fall-through logic won't behave as we want it
to here - a rate-limiting rule will only make processing abort for
a request when the rate limit is tripped.

the secret for x-forwarded-for enablement is generated by credhub
and woven through the creation pipeline. this will presumably be
needed by the cdn broker to eventually be able to set the
appropriate header. until that is done, the waf should continue
to work as it always previously has, not obeying x-forwarded-for,
so there's no race to set up both ends of this at the same time.